@@ -1,86 +1,48 @@
 {
-  "name": "@librocco/ui",
-  "version": "1.0.0",
-  "private": true,
-  "main": "dist/index.js",
-  "exports": {
-    "./dist/style.css": {
-      "import": "./dist/style.css"
-    },
-    ".": {
-      "import": "./dist/index.es.js",
-      "require": "./dist/index.js",
-      "default": "./dist.index.es.js"
-    }
-  },
-  "types": "dist/index.d.ts",
-  "peerDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "scripts": {
-    "build": "vite build && tsc --project tsconfig.build.json",
-    "lint": "eslint .",
-    "lint:strict": "eslint . --max-warnings=0",
-    "typecheck": "tsc --noEmit",
-    "test": "jest --watch",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@librocco/shared": "workspace:*",
-    "@babel/core": "7.x",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-typescript": "^7.14.5",
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-interactions": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
-    "@storybook/addon-postcss": "~2.0.0",
-    "@storybook/react": "^6.4.22",
-    "@storybook/testing-library": "^0.0.11",
-    "@testing-library/dom": "^8.13.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.1.1",
-    "@testing-library/user-event": "^14.1.1",
-    "@types/jest": "^27.4.1",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@types/testing-library__jest-dom": "^5.14.3",
-    "@typescript-eslint/eslint-plugin": "^5.20.0",
-    "@typescript-eslint/parser": "^5.20.0",
-    "@vitejs/plugin-react": "^1.3.1",
-    "autoprefixer": "~10.4.7",
-    "babel-loader": "^8.2.5",
-    "eslint": "^8.13.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-import-resolver-typescript": "^2.7.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
-    "eslint-plugin-unused-imports": "^2.0.0",
-    "jest": "^27.5.1",
-    "postcss": "~8.4.13",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "tailwindcss": "~3.0.24",
-    "typescript": "4.6.3",
-    "vite": "^2.9.9",
-    "@svgr/webpack": "5.5",
-    "vite-plugin-svgr": "~2.1.0",
-    "storybook-addon-pseudo-states": "~1.9.0"
-  }
+	"name": "@librocco/ui",
+	"version": "0.0.1",
+	"main": "./dist/index.js",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "svelte-kit package",
+		"preview": "vite preview",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
+		"format": "prettier --write --plugin-search-dir=. .",
+		"test": "vitest",
+		"storybook": "start-storybook -p 6006",
+		"build-storybook": "build-storybook"
+	},
+	"devDependencies": {
+		"@storybook/addon-actions": "^6.4.22",
+		"@storybook/addon-essentials": "^6.4.22",
+		"@storybook/addon-links": "^6.4.22",
+		"@storybook/addon-postcss": "^2.0.0",
+		"@storybook/addon-svelte-csf": "^2.0.3",
+		"@storybook/svelte": "^6.4.22",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
+		"@testing-library/svelte": "~3.1.3",
+		"@typescript-eslint/eslint-plugin": "^5.27.0",
+		"@typescript-eslint/parser": "^5.27.0",
+		"eslint": "^8.16.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"postcss": "~8.4.13",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"svelte": "^3.44.0",
+		"svelte-check": "^2.7.1",
+		"svelte-loader": "^3.1.2",
+		"svelte-preprocess": "^4.10.6",
+		"tailwindcss": "~3.0.24",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.4",
+		"vite": "^3.0.0",
+		"vitest": "~0.21.1",
+		"vitest-svelte-kit": "0.0.6",
+		"svelte2tsx": "~0.5.13"
+	},
+	"type": "module"
 }
