@@ -1,48 +1,9 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import svgr from "vite-plugin-svgr";
-import path from "path";
+import { UserConfig } from 'vite';
 
-const svgrOptions = require("../scaffold/svgrOptions");
+import { svelte as sveltekit } from '@sveltejs/kit/dist/vite';
 
-export default defineConfig({
-  resolve: {
-    alias: {
-      "@assets": path.resolve(__dirname, "src", "assets"),
-    },
-  },
-  build: {
-    lib: {
-      /** Maybe rename this to @<project-name>/ui */
-      name: "@librocco/ui",
-      entry: path.join(__dirname, "src", "index.ts"),
+const config: UserConfig = {
+	plugins: [sveltekit()]
+};
 
-      // Using this to build both the 'cjs' ans 'es' modules
-      // The appropriate entry points are defined in 'package.json'
-      fileName: (fmt) => (fmt === "es" ? "index.es.js" : "index.js"),
-      formats: ["es", "cjs"],
-    },
-    rollupOptions: {
-      // React should be used in the main app, and not bundled in
-      // as multiple instances of React can cause a bunch of problems
-      external: ["react", "react-dom"],
-      output: {
-        exports: "named",
-
-        // Global vars to use in UMD build for externalized deps
-        globals: {
-          react: "React",
-          "react-dom": "ReactDOM",
-        },
-      },
-    },
-    outDir: "dist",
-  },
-  plugins: [
-    react(),
-    svgr({
-      exportAsDefault: true,
-      svgrOptions,
-    }),
-  ],
-});
+export default config;
