@@ -5,7 +5,7 @@ import { createRoot } from "react-dom/client";
 import "./main.css";
 // Import ui .css file to make sure all of the tailwind classes, used in 'ui' package are available
 // in the final app (and not purged if the app src files don't use them)
-import "@librocco/ui/dist/style.css";
+import "@librocco/ui-react/dist/style.css";
 
 import App from "./App";
 
